requirements.txt
pillow==6.1.0
scipy==1.1.0
numpy==1.17.0
matplotlib==3.1.1
torch>=1.1.0
torchvision==0.4.0
scikit-image==0.15.0
flask ==1.1.1
gunicorn == 19.9.0
google-cloud-storage==1.18.0